Refactor with AuthProviderTypes, refactor AuthProviderType variable names
